feat: divide group member role
