+++ assets/i18n/strings_ko.i18n.json
@@ -66,6 +66,10 @@
     "manage": {
       "header": "그룹 관리",
       "back": "돌아가기",
+      "deleteConfirmationTitle": "정말 그룹을 삭제하시겠습니까?",
+      "deleteConfirmationMessage": "기존에 본 그룹 명의로 작성된 공지에는 영향을 끼치지 않습니다.",
+      "leaveConfirmationTitle": "정말 그룹을 나가시겠습니까?",
+      "leaveConfirmationMessage": "기존에 본 그룹 명의로 작성된 공지에는 영향을 끼치지 않습니다.",
       "change": "변경",
       "profileImage": "그룹 프로필 사진 변경",
       "name": {
@@ -79,11 +83,13 @@
       "notionLink": {
         "header": "노션 페이지 링크 변경",
         "hintText": "노션 링크 입력",
-        "loading": "노션 불러오는 중"
+        "loading": "노션 불러오는 중",
+        "error": "노션 링크를 불러오는데 실패했습니다"
       },
       "invite": {
         "header": "초대 링크 생성",
-        "invitationLink": "초대 링크"
+        "invitationLink": "초대 링크",
+        "copy": "초대 링크 복사하기"
       },
       "member": {
         "header": "멤버 관리"
@@ -96,7 +102,7 @@
         "role": "역할",
         "admin": "관리자",
         "manager": "매니저",
-        "user": "일반"
+        "member": "멤버"
       },
       "banish": "추방"
     },
