+++ assets/i18n/strings_ru.i18n.json
@@ -66,6 +66,10 @@
     "manage": {
       "header": "Управление группами",
       "back": "Назад",
+      "deleteConfirmationTitle": "Вы уверены, что хотите удалить группу?",
+      "deleteConfirmationMessage": "Это не повлияет на ранее опубликованные уведомления под этим именем группы.",
+      "leaveConfirmationTitle": "Вы уверены, что хотите покинуть группу?",
+      "leaveConfirmationMessage": "Это не повлияет на ранее опубликованные уведомления под этим именем группы.",
       "change": "изменение",
       "profileImage": "Изменить изображение профиля группы",
       "name": {
@@ -79,11 +83,13 @@
       "notionLink": {
         "header": "Обновить ссылку на страницу Notion",
         "hintText": "Введите ссылку на Notion",
-        "loading": "Загрузка Notion"
+        "loading": "Загрузка Notion",
+        "error": "Не удалось загрузить ссылку на Notion"
       },
       "invite": {
         "header": "Создать ссылку для приглашения",
-        "invitationLink": "Ссылка для приглашения"
+        "invitationLink": "Ссылка для приглашения",
+        "copy": "Копировать ссылку на приглашение"
       },
       "member": {
         "header": "Управление участниками"
@@ -96,7 +102,7 @@
         "role": "Роль",
         "admin": "Администратор",
         "manager": "Менеджер",
-        "user": "Пользователь"
+        "member": "член"
       },
       "banish": "Исключить"
     },
